dsl + pagination changes
