+++ examples/demo/src/models/users.rs
@@ -60,7 +60,7 @@ impl Authenticable for super::_entities::users::Model {
     async fn find_by_api_key(db: &DatabaseConnection, api_key: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
             .filter(
-                query::dsl::condition()
+                query::condition()
                     .eq(users::Column::ApiKey, api_key)
                     .build(),
             )
@@ -82,11 +82,7 @@ impl super::_entities::users::Model {
     /// When could not find user by the given token or DB query error
     pub async fn find_by_email(db: &DatabaseConnection, email: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
-            .filter(
-                query::dsl::condition()
-                    .eq(users::Column::Email, email)
-                    .build(),
-            )
+            .filter(query::condition().eq(users::Column::Email, email).build())
             .one(db)
             .await?;
         user.ok_or_else(|| ModelError::EntityNotFound)
@@ -103,7 +99,7 @@ impl super::_entities::users::Model {
     ) -> ModelResult<Self> {
         let user = users::Entity::find()
             .filter(
-                query::dsl::condition()
+                query::condition()
                     .eq(users::Column::EmailVerificationToken, token)
                     .build(),
             )
@@ -120,7 +116,7 @@ impl super::_entities::users::Model {
     pub async fn find_by_reset_token(db: &DatabaseConnection, token: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
             .filter(
-                query::dsl::condition()
+                query::condition()
                     .eq(users::Column::ResetToken, token)
                     .build(),
             )
@@ -138,7 +134,7 @@ impl super::_entities::users::Model {
         let parse_uuid = Uuid::parse_str(pid).map_err(|e| ModelError::Any(e.into()))?;
         let user = users::Entity::find()
             .filter(
-                query::dsl::condition()
+                query::condition()
                     .eq(users::Column::Pid, parse_uuid)
                     .build(),
             )
@@ -155,7 +151,7 @@ impl super::_entities::users::Model {
     pub async fn find_by_api_key(db: &DatabaseConnection, api_key: &str) -> ModelResult<Self> {
         let user = users::Entity::find()
             .filter(
-                query::dsl::condition()
+                query::condition()
                     .eq(users::Column::ApiKey, api_key)
                     .build(),
             )
@@ -184,7 +180,7 @@ impl super::_entities::users::Model {
 
         if users::Entity::find()
             .filter(
-                query::dsl::condition()
+                query::condition()
                     .eq(users::Column::Email, &params.email)
                     .build(),
             )
