+++ src/lib.rs
@@ -47,7 +47,6 @@ pub mod cli;
 
 pub mod auth;
 pub mod boot;
-pub mod concern;
 pub mod config;
 pub mod controller;
 pub mod environment;
